Changed made in coins.js, backgroundDesign,js, movement.js and also commented on the uncomment part from the last commit
In the coins.js I have fixed the collision. I saw in isCoinStartYIsInPlayer function I did not use playerEndY at all hence there was an issue with the Y collision. I also added 2 more functions to detect of player is in coin's rectangle or not. Previously I was only checking if coin was in player's rectangle was or not. I know they sound the same but player changes it's width and height according to it's movement direction hence I needed to add two more functions and used them inside the for-loop where I check for collision.

backgroundDesing.js saw a resize() getting added just for the sake of having the timer label and score label stay on the screen even if the user resizes the window.

movement.js had the animation part on in all the move codes and got the css update code commented out. You can use either one of them if you like. Just have to make sure you change the speed variable accordingly.

Added some comments and also added a code to store score in local storage so that user can see how many scores they achieved in the last scene.